# frontend/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "http://restio-openapi-app:8080", "description": "Generated server url" }],
  "paths": {
    "/api/v1/restaurants/{restaurantId}/sync/events": {
      "post": {
        "tags": ["sync"],
        "summary": "Upload a batch of events; idempotent by eventId, one result per event",
        "operationId": "upload",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UploadEventsRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/UploadEventsResponse" } }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/devices/{deviceId}/deactivate": {
      "post": {
        "tags": ["devices"],
        "summary": "Unlink a device for good; its token and PIN logins stop working",
        "operationId": "deactivate",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/DeviceResponse" } } }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/devices/pairings": {
      "post": {
        "tags": ["devices"],
        "summary": "Create a one-time pairing code (QR) to link a new device",
        "operationId": "createPairing",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePairingRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/PairingResponse" } } }
          }
        }
      }
    },
    "/api/v1/devices/pair": {
      "post": {
        "tags": ["devices"],
        "summary": "Redeem a pairing code from the new device; returns its token once",
        "operationId": "pair",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PairRequest" } }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/PairResponse" } } }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Exchange a refresh token for a new token pair",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } }
          }
        }
      }
    },
    "/api/v1/auth/pin-login": {
      "post": {
        "tags": ["auth"],
        "summary": "Log in with a PIN from a registered device (shift session, no refresh)",
        "operationId": "pinLogin",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PinLoginRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Revoke a refresh token",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Log in with email and password",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } }
          }
        }
      }
    },
    "/api/v1/auth/device-login": {
      "post": {
        "tags": ["auth"],
        "summary": "Session of a linked device itself, for synchronisation (no user)",
        "operationId": "deviceLogin",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DeviceLoginRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } }
          }
        }
      }
    },
    "/api/v1/auth/change-pin": {
      "post": {
        "tags": ["auth"],
        "summary": "Change the PIN, proving identity with the current PIN or password",
        "operationId": "changePin",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ChangePinRequest" } }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Change the password; closes every other session",
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/restaurants/{restaurantId}/sync/snapshot": {
      "get": {
        "tags": ["sync"],
        "summary": "Master data a device needs to work offline",
        "operationId": "snapshot",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/SnapshotResponse" } } }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/devices": {
      "get": {
        "tags": ["devices"],
        "summary": "Devices of the restaurant, active and unlinked",
        "operationId": "list",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/DeviceResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permissions": {
      "get": {
        "tags": ["auth"],
        "summary": "Catalog of permissions",
        "operationId": "list_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "array", "items": { "type": "string" } } } }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Current user, restaurants and permissions",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/MeResponse" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EventDto": {
        "type": "object",
        "properties": {
          "eventId": { "type": "string", "format": "uuid" },
          "type": { "type": "string", "maxLength": 100, "minLength": 0 },
          "aggregateId": { "type": "string", "format": "uuid" },
          "restaurantId": { "type": "integer", "format": "int64" },
          "deviceId": { "type": "string", "format": "uuid" },
          "userId": { "type": "integer", "format": "int64" },
          "schemaVersion": { "type": "integer", "format": "int32" },
          "payload": { "$ref": "#/components/schemas/JsonNode" },
          "hlc": { "type": "string", "maxLength": 64, "minLength": 0 },
          "epoch": { "type": "integer", "format": "int64" },
          "seq": { "type": "integer", "format": "int64" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "aggregateId",
          "createdAt",
          "deviceId",
          "epoch",
          "eventId",
          "hlc",
          "payload",
          "restaurantId",
          "schemaVersion",
          "type"
        ]
      },
      "JsonNode": {},
      "UploadEventsRequest": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EventDto" },
            "maxItems": 500,
            "minItems": 0
          }
        },
        "required": ["events"]
      },
      "EventResultDto": {
        "type": "object",
        "properties": {
          "eventId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["APPLIED", "DUPLICATE", "REJECTED"] }
        }
      },
      "UploadEventsResponse": {
        "type": "object",
        "properties": {
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/EventResultDto" } }
        }
      },
      "DeviceResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "uuid": { "type": "string", "format": "uuid" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["POS", "KDS", "ADMIN"] },
          "active": { "type": "boolean" },
          "lastSeenAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreatePairingRequest": {
        "type": "object",
        "properties": {
          "deviceName": { "type": "string", "maxLength": 100, "minLength": 0 },
          "deviceType": { "type": "string", "enum": ["POS", "KDS", "ADMIN"] }
        },
        "required": ["deviceName", "deviceType"]
      },
      "PairingResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
      "PairRequest": {
        "type": "object",
        "properties": { "code": { "type": "string", "maxLength": 20, "minLength": 0 } },
        "required": ["code"]
      },
      "PairResponse": {
        "type": "object",
        "properties": {
          "deviceId": { "type": "integer", "format": "int64" },
          "deviceUuid": { "type": "string", "format": "uuid" },
          "deviceCode": { "type": "string" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["POS", "KDS", "ADMIN"] },
          "restaurantId": { "type": "integer", "format": "int64" },
          "deviceToken": { "type": "string" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "properties": { "refreshToken": { "type": "string", "minLength": 1 } },
        "required": ["refreshToken"]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn": { "type": "integer", "format": "int64" }
        }
      },
      "PinLoginRequest": {
        "type": "object",
        "properties": {
          "deviceId": { "type": "integer", "format": "int64" },
          "deviceToken": { "type": "string", "minLength": 1 },
          "pin": { "type": "string", "minLength": 1, "pattern": "\\d{4,6}" }
        },
        "required": ["deviceId", "deviceToken", "pin"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "minLength": 1 },
          "password": { "type": "string", "minLength": 1 }
        },
        "required": ["email", "password"]
      },
      "DeviceLoginRequest": {
        "type": "object",
        "properties": {
          "deviceId": { "type": "integer", "format": "int64" },
          "deviceToken": { "type": "string", "minLength": 1 }
        },
        "required": ["deviceId", "deviceToken"]
      },
      "ChangePinRequest": {
        "type": "object",
        "properties": {
          "currentSecret": { "type": "string", "minLength": 1 },
          "newPin": { "type": "string", "minLength": 1, "pattern": "\\d{4,6}" }
        },
        "required": ["currentSecret", "newPin"]
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": { "type": "string", "minLength": 1 },
          "newPassword": { "type": "string", "maxLength": 72, "minLength": 8 }
        },
        "required": ["currentPassword", "newPassword"]
      },
      "PosUserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "displayName": { "type": "string" },
          "pinHash": { "type": "string" },
          "roles": { "type": "array", "items": { "type": "string" }, "uniqueItems": true },
          "permissions": { "type": "array", "items": { "type": "string" }, "uniqueItems": true }
        }
      },
      "SnapshotResponse": {
        "type": "object",
        "properties": {
          "generatedAt": { "type": "string", "format": "date-time" },
          "users": { "type": "array", "items": { "$ref": "#/components/schemas/PosUserDto" } }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "email": { "type": "string" },
          "displayName": { "type": "string" },
          "restaurantIds": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" },
            "uniqueItems": true
          },
          "roles": { "type": "array", "items": { "type": "string" }, "uniqueItems": true },
          "permissions": { "type": "array", "items": { "type": "string" }, "uniqueItems": true },
          "deviceId": { "type": "integer", "format": "int64" }
        }
      }
    }
  }
}